move logos into image
