+++ src/elements/image/logoSmall.svelte
