+++ src/elements/navigation/header.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-	import LogoSmall from '../logoSmall.svelte';
+	import LogoSmall from 'elements/image/logoSmall.svelte';
 	import { fade } from 'svelte/transition';
 
 	export let menu: [string, string][];
